// src/types/apply/general-state.ts
export type GeneralState = {
  name: string
  dob: Date
  gender: 'Nam' | 'Nữ'
  highSchool: string
  major: string
  phoneNumber: string
  email: string
  facebook: string
  avatar: string
  extracurriculars: string
  strengthsWeaknesses: string
  uniqueness: string
  futurePlans: string
  photoThoughts: string
}